fix: remove support for load viz on fixture
