+++ commands/serve/web/eRender.html
@@ -8,7 +8,6 @@
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:light" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:semibold" rel="stylesheet" />
-    <script src="/stardust.js"></script>
     <% for (var stylesheet in htmlWebpackPlugin.options.stylesheets) { %>
     <script type="text/javascript" src="<%= htmlWebpackPlugin.options.stylesheets[stylesheet] %>"></script>
     <% } %> <% for (var script in htmlWebpackPlugin.options.scripts) { %>
