+++ test/component/barchart/sn-barchart/src/data.js
