+++ test/component/barchart/sn-barchart/src/index.js
