+++ test/component/barchart/sn-barchart/src/object-properties.js
