+++ test/component/barchart/sn-barchart/src/pic-definition.js
